Fix Phase 2/3 BER Calculation, Change Plots
This commit fixes BER Calculation as it was originally only showing
total amount of bit errors instead of bit error rate.

This commit also fixes error rate array from a NxN matrix to a Nx1 array

This commit also modifies the plot of the modulated signal to use
spectrum analysis.